+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.5.0
 - Added "gitGrace.branch" command.
 - Added "gitGrace.checkout" command.
+- Added "gitGrace.stashPopLatest" command.
 - Fixed "gitGrance.master" and "gitGrace.deleteMergedBranches" commands so they stop asking for fast forward.
 
 ### 1.4.1